fix(storyboard): split storyboardPromptCompiler out of storyboardPlan.ts to stay under the filesize gate
PR #492 left storyboardPlan.ts at exactly 800 lines and PR #494 added 18 more.
Each was green on its own branch; merged together they take main's
check:filesize over the 800-line ceiling.

Extract the plan-to-prompt compilation layer (pure functions: anchor sheet
prompt, shot prompt, keyframe prompt, the shared anchor prompt bits, the
prompt-derived visual reference order, and the isVisualAnchor predicate they
share) into storyboardPromptCompiler.ts. Behavior is unchanged: the block is
moved verbatim, storyboardPlan.ts imports it, and the three external consumers
import the two public symbols from their new home instead of a re-export shim.

storyboardPlan.ts 818 -> 685 lines; storyboardPromptCompiler.ts 147 lines.
The back-reference is type-only, so no runtime cycle (check:boundaries green).
